requirements.txt
click==8.0.4
colorama==0.4.4
feedparser==6.0.8
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
packaging==21.3
psycopg2-binary==2.9.3
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
schedule==1.1.0
sgmllib3k==1.0.0
six==1.16.0
SQLAlchemy==1.4.32
SQLAlchemy-Utils==0.38.2
typing_extensions==4.1.1
Werkzeug==2.0.3
zipp==3.7.0
